add click

+++ ristorino-backend/src/main/java/ubp/das/ristorinobackend/entity/SucursalRestauranteId.java
@@ -13,5 +13,5 @@ public class SucursalRestauranteId implements Serializable {
     private Long nroRestaurante;
 
     @Column(name = "nro_sucursal")
-    private Integer nroSucursal;
+    private Short nroSucursal;
 }
